+++ src/common/utils/index.ts
@@ -5,4 +5,4 @@ export * from './module.decorator';
 export * from './dependencies.decorator';
 export * from './inject.decorator';
 export { Component as Middleware } from './component.decorator';
-export * from './provide-values.util';
+export * from './merge-with-values.util';