+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0-RC7 (08.04.2017)
+
+- MiddlewareBuilder: `use()` deprecated, use `apply()` instead
+- MiddlewareBuilder: new `apply()` method
+
 ## 1.0.0-RC4 (08.04.2017)
 
 - Support for @Post, @Get, @Delete, @Put, @All decorators